refactor: batch에서 사용하는 jpa repository, entity 클래스를 상위 common 패키지 하위로 이동
