+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/common/entity/ShortTermWeatherForecast.java
@@ -1,4 +1,4 @@
-package com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.writer.entity;
+package com.ImSnacks.NyeoreumnagiBatch.common.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
