+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/common/entity/ShortTermWeatherForecastId.java
@@ -1,4 +1,4 @@
-package com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.writer.entity;
+package com.ImSnacks.NyeoreumnagiBatch.common.entity;
 
 import lombok.*;
