+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/common/entity/UniqueNxNy.java
@@ -1,4 +1,4 @@
-package com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.reader.entity;
+package com.ImSnacks.NyeoreumnagiBatch.common.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
