+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/common/repository/ShortTermWeatherForecastRepository.java
@@ -1,0 +1,9 @@
+package com.ImSnacks.NyeoreumnagiBatch.common.repository;
+
+import com.ImSnacks.NyeoreumnagiBatch.common.entity.ShortTermWeatherForecast;
+import com.ImSnacks.NyeoreumnagiBatch.common.entity.ShortTermWeatherForecastId;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ShortTermWeatherForecastRepository extends JpaRepository<ShortTermWeatherForecast, ShortTermWeatherForecastId> {
+
+}
