--- batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/shortTermWeatherForecast/writer/repository/ShortTermWeatherForecastRepository.java
@@ -1,9 +1,0 @@
-package com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.writer.repository;
-
-import com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.writer.entity.ShortTermWeatherForecast;
-import com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.writer.entity.ShortTermWeatherForecastId;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ShortTermWeatherForecastRepository extends JpaRepository<ShortTermWeatherForecast, ShortTermWeatherForecastId> {
-
-}
