+++ batch/src/test/java/com/ImSnacks/NyeoreumnagiBatch/filterTest/TemperatureFilterTest.java
@@ -4,7 +4,7 @@ import com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.processor.dto.Vil
 import com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.processor.utils.weatherRiskFilter.TemperatureFilter;
 import com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.processor.utils.weatherRiskFilter.WeatherRiskFilter;
 import com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.writer.dto.ShortTermWeatherDto;
-import com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.writer.entity.WeatherRiskType;
+import com.ImSnacks.NyeoreumnagiBatch.common.entity.WeatherRiskType;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
